refactor(frontend): extract log row renderers
